requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
boto3==1.21.2
botocore==1.24.2
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools==5.0.0
certifi==2021.10.8
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1636046173594/work
charset-normalizer==2.0.12
click==8.0.3
filelock==3.5.0
flatbuffers==2.0
frozenlist==1.3.0
fsspec==2022.1.0
future @ file:///Users/runner/miniforge3/conda-bld/future_1635819640244/work
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio @ file:///Users/runner/miniforge3/conda-bld/grpcio_1639715812589/work
h5py @ file:///Users/runner/miniforge3/conda-bld/h5py_1637964045648/work
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.0
jmespath==0.10.0
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
multidict==6.0.2
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1640083125610/work
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1644499459631/work
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyDeprecate==0.3.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytorch-transformers==1.2.0
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1636140145796/work
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.1
sacremoses==0.0.47
semantic-version==2.9.0
sentencepiece==0.1.96
setuptools-rust==1.1.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-macos==2.8.0
tensorflow-metal==0.3.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tokenizers==0.11.4
torch @ file:///Users/runner/miniforge3/conda-bld/pytorch-recipe_1643987874214/work
torchmetrics==0.7.2
torchvision==0.2.2
tqdm==4.62.3
transformers @ git+https://github.com/huggingface/transformers@2d02f7b29b8c18a68260378246f9e26621021c30
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1644850595256/work
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
yarl==1.7.2
zipp==3.7.0
